Tentative start on web scraping SINoALICE db
Currently using Puppeteer to load the website as it uses client sided rendering. Currently, it is possible to use puppeteer to load the web page and click buttons. The nightmares.js file uses puppeteer to click the fields button of the website and emulates button clicks to select the desired fields that we want to scrape from the website. The next step is to store the data into JSON form and return it. The example.png is an screenshot taken using puppeteer of how the website looks after the desired fields have been selected. It was also used as part of the debugging process to confirm whether the code was working properly