Interim Commit. Refactoring to use SINoALICE db API to get nightmares instead of web scraping
Didn't realise it was possible to use their API. Initially thought of using it before I started the web scraper code, but didn't work because I didn't specify CORS in the fetch request. Just thought of retrying it, and it turns out that it's actually possible if CORS is specified in the fetch. Currrently refactoring to use db API for nightmare list, although may still need to do some web scraping, since the EN skill description isn't included with the nightmare list for some reason despite the JP description being included