Added ability to select and add nightmares to the timeline
This was by far the most difficult function to implement. The timeline and selectedNightmare list was always reset to the default state because the callback assigned to the image list was always using the old state values at the time of rendering the image list. Spent about 7 hours trying to figure out why set state didn't work only to find out that it was working, but it was updating using based off of the default initial state, hence why it seemed as if state wasn't being updated correctly. This was worked aroud using stateRefs, which allowed the image list a way to access the latest state without remapping the callbacks for every image in the list (avoiding re-render of image list component)